[Feature #161214014] add endpoint to get all sales records
